prepared some structure of what I'd like to do now
So if we could start tracking without reading from files it means that
we can prepare the buffer before tracking during the separate steps of the algorithm and then we need still to replace the fb_next and read_frame and so on by fb_load or something like this? or maybe all is not necessary - we like give tracker the prio from the previous chunk and let it run like for the first time every time, not continuosly and not membering anything backwards. so we are now will be limited to run tracking only to the size of the buffer itself. not a nice one.
if we want to have 10 frames and track using our 4-step algorithm, then we have to create functions that are reading from some memory, fill and advance the buffer in these 10 frames.